requirements.txt
boto3
botocore
pypdfium2==4.30.0
faiss-cpu==1.8.0.post1
PyMuPDF==1.24.10
duckduckgo_search
langchain==0.2.16
langchain-community==0.2.17
langchain-aws==0.1.18
tabula
tqdm
pillow
matplotlib
langchain_postgres
python-dotenv
pgvector
psycopg-binary
mistralai
scipy
Flask==3.0.0
python-pptx==0.6.21
numpy==1.24.3
